requirements.txt
plotly-express==0.4.1
clean-text==0.6.0
textblob==0.17.1
altair==4.2.0
protobuf==3.20.3
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.57.0
pandas==1.5.2
streamlit==1.17.0
Python==3.11.8
pandas==2.2.1
google-api-python-client==2.122.0
oauth2client==4.1.3
python-csv==0.0.13
python-time==0.3.0
DateTime==5.5
isodate==0.6.1
